requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.5
django-js-asset==1.2.2
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
